+++ example/myapp/components/coffee/search/search.py
@@ -1,9 +1,9 @@
-from django_components import component
+from django_components import register
 
 from livecomponents import CallContext, StatelessLiveComponent, command
 
 
-@component.register("coffee/search")
+@register("coffee/search")
 class SearchComponent(StatelessLiveComponent):
     template_name = "coffee/search/search.html"