+++ example/myapp/components/emailsender/emailsender.py
@@ -1,9 +1,9 @@
-from django_components import component
+from django_components import register
 
 from livecomponents import CallContext, StatelessLiveComponent, command
 
 
-@component.register("emailsender")
+@register("emailsender")
 class EmailSender(StatelessLiveComponent):
     template_name = "emailsender/emailsender.html"